+++ packages/scms-core/src/components/index.ts
@@ -17,8 +17,6 @@ export * from './SystemAdminBadge.js';
 export * from './ThemeSwitcher.js';
 export * from './UserCard.js';
 export * from './VersionsListing.js';
-
-export * from './dialogs/index.js';
 export * from './layout/index.js';
 export * from './navigation/index.js';
 export * from './upload/index.js';
