+++ packages/scms-core/src/components/ui/dialogs/SimpleDialog.tsx
@@ -8,9 +8,9 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
-} from '../ui/dialog.js';
-import { Button, type ButtonProps } from '../ui/button.js';
-import { cn } from '../../utils/cn.js';
+} from '../dialog.js';
+import { Button, type ButtonProps } from '../button.js';
+import { cn } from '../../../utils/cn.js';
 import type { DialogContentProps } from '@radix-ui/react-dialog';
 
 export interface DialogButton {
