+++ packages/scms-core/src/components/ui/dialogs/index.ts
@@ -1,2 +1,3 @@
 export * from './ResumeDraftWorkDialog.js';
 export * from './SimpleDialog.js';
+export * from './types.js';
